feat: Extended autoplay options (#116)
This change adds a couple of string options for the `autoplay` attribute/`autoPlay` prop. The exiting boolean-attribute nature of `autoplay`/`autoPlay` lives on.
Specifically:
- `muted` - will mute the player and try autoplaying. If autoplaying fails, muted state is restored and player continues as a standard click-to-play player.
- `any` - will try playing with the current options, if it fails it'll behave like `muted` above.
  Happy to get alternatives to the name `any`.

In addition to the above, this switches from relying on native autoplay attribute behavior to calling `play` manually. In terms of behavior, it should be backwards compatible, but will allow us to better control this behavior going forward.

Example usage:
WC
```html
<!-- no autoplay -->
<mux-player ></mux-player>
<!-- autoplay muted, likely to work -->
<mux-player autoplay="muted"></mux-player>
<!-- autoplay with sound, likely to fail -->
<mux-player autoplay></mux-player>
```
React:
```jsx
// autoplay with sound, if possible, then muted if possible
<MuxPlayer autoPlay="any" />
```

Co-authored-by: Christian Pillsbury <cjpillsbury@gmail.com>