feat(mux-player): move captions up when controls are showing
Currently, when using native captions, when the controls are showing
they obscure the captions.

We should move captions up when the controls are showing. This means
that when the player is paused or the user is active we should move the
captions up.

Currently, this only accounts for captions that are at the bottom of the
video viewport. Eventually, we may want to consider moving captions that
are at the top or those that are in the middle as we have controls in
those places as well.