chore: update builds to meet devs where they are
This means that we now output commonjs and esm bundles that target
native node (for SSR) or bundler users. These bundles externalize all
dependencies where it makes sense to promote re-usability.

The react modules currently only produce a commonjs build as that is
what React itself produces.

We also output an IIFE bundle names `<package-name>.js` for `<script>`
users. In addition, an ESM bundle that bundles all dependencies in is
available as `<package-name>.mjs` for `<script type="module">` users.

There is a metafile generated for each bundle which can be inspected via
bundle-buddy or esbuild-visualizer.

All the polyfills are now in the shared/polyfills package.

Tests are also updated to use import-maps with a description of the
changes in the shared/test-esm-exports local package.