fix: ignore Safari for captions movement. (#404)
Safari has a bug where caption cues are duplicated when the original cue
from an HLS stream with native playback is modified
(https://bugs.webkit.org/show_bug.cgi?id=240549).

This disables the caption movement behavior, which moves captions up so
that the player controls don't obscure the captions.

Unfortunately, the pseudo elements -webkit-media-text-track-display
isn't working right now as an alternative method. The other solutions
are rendering captions ourselves or switching to MSE in Safari but we
aren't ready to do either of these yet.