chore: add scrollable options in nextjs/MuxPlayer
Sets the height of the options section to be 100vh minus the player
height and everything else.

Player height defaults to 450 and will grab the first height that's set
when loaded. Will update this if the player size is changed via the
"width cutoffs for reponsinve player chrome/ui" option.