feat(mux-player): add storyboard-src attriubte
Before this commit, it was not possible to set the source URL for the
player's storyboard asset to anything other than what was derived from
the `sourceId`.

After this commit, a user can set `storyboard-src` to point to any URL
that contains the storyboard. Then, the `storyboard()` getter will return
`this.storyboard()` instead of trying to derive the URL from the
`sourceId`.`