Correct package dependencies (#666)
This PR adds all transitive dependencies. As you can select pnpm mode in
node linker config (which introduced in yarn v3), I recommend that you
update to yarn v3 whenever possible and use that mode. It's the only
non-pnpm package manager option that can prevent problems with
transitive dependencies out of the box. And it's much more stable than
pnpm.

There is already a not bad explanation, so just quote[^1] that:

> ### Why not NPM or YARN (v1)?
> NPM uses flattened node_modules. The flattened dependency results in
many issues, such as:
The algorithm for flattening a dependency tree is complex
Some of the packages have to be duplicated inside another project's
node_modules folder.
Modules have access to packages they do not depend on
Let's explain this with a basic example. One of your projects needs an
express module, and express has a dependency package named "debug".
Your code can require('debug'), even if you do not depend on it
explicitly in the package.json file.

[^1]:
https://dev.to/ayoub3bidi/lets-settle-things-out-2-npm-vs-yarn-vs-pnpm-5e04

---------

Signed-off-by: Sora Morimoto <sora@morimoto.io>