fix: force theme to be ltr direction (#670)
if the player is set to rtl direction, some functionality breaks, so, we
want to force it to always be ltr

If we set it on the :host, then if someone sets direction on mux-player
itself, it'll get overridden. So, instead, to force it to always be ltr,
we set it on the media-theme element itself.