fix: use webkit pseudo element for captions movement, where available (#674)
While a single solution for all browsers is nice, we've had some reports
of captions briefly showing up at the top of the video display area,
which is unexpected. Therefore, we should switch back to the pseudo
element technique for Chrome and Safari while keeping the programmatic
line technique in use for Firefox, since it doesn't have anything else
available.

This uses a transform on the text track container pseudo element and
will transition it at 0.433s so that the transition is nice and smooth.
This number was borrowed from the WebVTT specification. Also, when
hiding, add a slight delay.

Fixes #660.