fix(playback-core): Apply pseudo-ended logic on seeked, even if paused. (#707)
NOTE: This is actually a bug in browsers based on a previously ambiguous
definition in the HTML spec. For context, see:
- https://github.com/video-dev/hls.js/issues/5168
- https://html.spec.whatwg.org/multipage/media.html#reaches-the-end
(referenced in the issue, above)

Additionally, I've updated the "pseudo-ended" logic to not apply when
the media element is looping, since (again, per spec), that media should
not ever "end".