chore: Prepare default theme from Mux Player for export (#668)
The goal of this PR is to package the existing default theme such that
it can be exported as it's own module. This can then be easily swapped
out in the future with a new theme.

Theme source files now reside in `src/themes` (themes based in this repo
and re-exported from media chrome). Build artefacts for themes are in
`dist/themes`.

---------

Co-authored-by: Gary Katsevman <git@gkatsev.com>
Co-authored-by: Christian Pillsbury <cpillsbury@mux.com>