feat: add poster slot for progressive enhancement (#747)
related PR https://github.com/muxinc/media-chrome/pull/721

this change makes it possible to set a poster slot. it results in
quicker loading of the thumbnail because the image url can be loaded in
parallel with the player JS bundle. this also allows users to better
make use of responsive images etc.

```html
<mux-player playback-id="VcmKA6aqzIzlg3MayLJDnbF55kX00mds028Z65QxvBYaA">
  <img slot="poster" src="https://image.mux.com/VcmKA6aqzIzlg3MayLJDnbF55kX00mds028Z65QxvBYaA/thumbnail.webp">
</mux-player>
```

this change also enables setting a poster for an audio player. by
default we don't set a poster src if the mux player is in audio mode but
the user can explicitly set the poster attribute and then the poster
image will show.

fixes #590