fix: Extra spacing underneath player (#779)
There's a few extra pixels of white space below a default mux player
integration. This is caused by the mux-theme element being `display:
inline-block` and poking out, even if the mux-player element itself is
display block. This is more obvious if you want to do curved borders:

<img width="925" alt="image"
src="https://github.com/muxinc/elements/assets/9952680/129eff41-04ab-41ac-bfa0-0afdcf02ca93">

This is fixable by setting `line-height: 0` on the mux element, but
feels a bit hacky and unintuitive. It's not obvious what the cause of
the extra space is from the developers perspective.

~~`display: block` fixes it, if that doesn't interfere with anything
else?~~

I've now internalised the `line-height: 0` fix which has the same
effect.