fix(playback-core): preload=none not respected for playlist stream info fetching with native safari playback. (#916)
## Background
In order to support for "media ui extensions" such as "stream type" in
our media elements across as many playback environments as possible, we
independently fetch and parse HLS playlists, even when playback of HLS
media is managed natively by the Safari `HTMLMediaElement`. We also aim
to respect developer preferences such as `preload="none"`. To accomplish
this, we previously relied on waiting for the `'loadstart'` event to be
dispatched before initiating our fetching logic for cases where
`preload="none"`. Unfortunately, per spec, `'loadstart'` will be
dispatched regardless of `preload` state. Both Chrome and Safari comply
with the specification (Firefox appears to deviate from the spec to get
closer to the "spirit of the law," but this also results in
cross-browser inconsistencies).

For the exact part of the HTML Living Standard that discusses this, see
Steps 7 & 8 of _The **resource selection algorithm**_ in [_§4.8.11.5
Loading the media resource_
](https://html.spec.whatwg.org/multipage/media.html#concept-media-load-algorithm)

To demonstrate this behavior, load [this
codesandbox](https://codesandbox.io/p/sandbox/media-loadstart-behavior-h9htzm)
in the respective browsers.

## The fix
To accomplish these goals but still attempt to retrieve and compute the
playlist+stream info as early as possible, we can instead listen for the
first `'play'` event or `'loadedmetadata'` event, whichever comes first.
We also need to make sure we actually set the underlying
`HTMLMediaElement`'s `preload` state before applying this logic (which
we were not previously doing.

fixes #875