fix(playback-core): pseudo-ended eval case where media is not attached. (#932)
This would definitely be an edge case, but is hypothetically possible
given the async aspects at play here. tl;dr - `buffered` time range
length may still be `0` when one of the several events that invoke
`maybeDispatchEndedCallback()` is dispatched. This ensures that we don't
end up in cases where we are assuming `buffered.length > 0` and also
ensures that we treat an empty buffered as under-determined for
"pseudo-ended" cases.