chore: replace Lerna w/ Release Please (#1036)
and use Turbo for task runner.

This change replaces Lerna with
https://github.com/googleapis/release-please for releasing our Mux
elements packages.

We started using Lerna right before it was deprecated and then saved and
taken over by nrwl https://github.com/lerna/lerna/issues/3121

Since then we haven't updated from v4, current version is v8, to avoid
breaking things.

With the creation of the media-elements monorepo and the player.style
monorepo we started using Release Please which turned out to be a good
experience. There is no local version bump needed and manual GH action
triggers etc. but keeps the good parts like use of conventional commits
and a confirmation step with an auto created release PR.

Canaries are still done via the publish.sh script but I'm happy to move
that over to use [wet-run](https://github.com/luwes/wet-run) if y'all
are on board. It will create provenance NPM versions and the canary
prerelease numbers will actually increment which is broken at the
moment. The SHA hash is not needed anymore because of the provenance
statements https://docs.npmjs.com/generating-provenance-statements