fix: bring back cast button for drm protected videos (#1137)
## Context
I have been writing documentation for building a custom Chromecast
sender and receiver for DRM playback. With the help of @luwes we
discovered that the Mux player sends everything you need for a custom
receiver to play DRM content. Unfortunately it's still not usable
because we always hide the cast button when a DRM token is provided.

This PR brings back the cast button for DRM content if you configure a
custom cast receiver.


## Testing

### Example code
```
    <mux-player
      id="player"
      playback-id="..."
      playback-token="..."
      drm-token=".."
      cast-receiver=""
    ></mux-player>
```

### In production 
Using the `<mux-player>` with a drm-token attribute will *hide* the cast
button
Using the `<mux-player>` with a drm-token attribute and a cast-receiver
attribute will *hide* the cast button

### In this PR
Using the `<mux-player>` with a drm-token attribute will *hide* the cast
button
Using the `<mux-player>` with a drm-token attribute and a cast-receiver
attribute will *show* the cast button