fix: Add MENU_ITEM constant for styling of menu items with CSS (#1212)
Currently menu items appear like they can be targeted and styled with
`mux-player::part(menu-item)` but this fails because the part
`menu-item` is not included in the list of parts to export.

Menu item markup

```html
<media-chrome-menu-item type="radio" role="menuitemradio" part="menu-item radio" value="1" aria-checked="false" tabindex="-1">...</media-chrome-menu-item>
```

Theme parts exported

```html
<media-theme defaultstreamtype="on-demand" defaultsubtitles="" forwardseekoffset="10" backwardseekoffset="10" exportparts="top, center, bottom, layer, media-layer, poster-layer, vertical-layer, centered-layer, gesture-layer, controller, button, range, display, control-bar, menu-button, menu, option, poster, live, play, pre-play, seek-backward, seek-forward, mute, captions, airplay, pip, fullscreen, cast, playback-rate, volume, time, title, audio-track, rendition" title="Test multiple captions and audio streams" videotitle="Test multiple captions and audio streams" style="--media-accent-color: #da2128;">
  ...
  </media-theme>
```

Also noticed `thumb` missing too, to allow for styling the thumb in the
volume and time ranges.

Could someone please point me in the right direction if this isn't the
place to suggest this. Thanks!

---------

Co-authored-by: Ronald Urbina <140639086+ronalduQualabs@users.noreply.github.com>