fix: Add display names to React Components (#1237)
This PR relates to an issue that happens commonly on next-video. This
does not fix the issue, but makes the warning message more clear.

Relates to:
- https://github.com/muxinc/next-video/issues/397
- https://github.com/muxinc/ce-la-react/issues/6

When rendering a MuxVideo element from `@mux/mux-video/react` we see
this message:
```
Check the render method of `i`. See https://react.dev/link/warning-keys for more information.
Each child in a list should have a unique "key" prop.
```

This PR changes the message to be:
```
Check the render method of `MuxVideo`. See https://react.dev/link/warning-keys for more information.
Each child in a list should have a unique "key" prop.
```

This also happens with `hls-video-element/react` and
`dash-video-element/react` (which already have the displayName).
You can test this by running the example in `next-video` (while [this
issue](https://github.com/muxinc/next-video/issues/397) is open) and
seeing the Next console logs.