fix: Properly setup DRM to support Airplay (#1245)
Closes [#183](https://github.com/muxinc/devextravaganza/issues/183)

I also reorganized the `onFpEncrypted` callback to get a better
understanding of the code and compare it with the [FairPlay Streaming
Overview](https://developer.apple.com/streaming/fps/FairPlayStreamingOverview.pdf).

The main issue here was that we were ignoring the Apple TV SPC when
generating the DRM license. This is reflected on the
`session.addEventListener('message', ...)` changes. We were listening
just once to this message, but more recent versions of Apple TV send new
messages when using AirPlay. Therefore I modified that listener to ask
for that license again and update the session for each new message.